fix: satisfy trace experiment audit
